+++ R/eem_correction.R
@@ -178,6 +178,7 @@ eem_remove_blank <- function(eem, blank = NA) {
 #'
 #' # Remove the scattering
 #' eem <- eem_remove_scattering(eem = eem, type = "raman", order = 1, width = 10)
+#' eem <- eem_remove_scattering(eem = eem, type = "rayleigh", order = 1, width = 10)
 #'
 #' plot(eem)